Backend/requirements.txt
blinker==1.6.3
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==3.0.0